refactor: major overhaul of project structure and entry points
- migrate core logic from src directory to lib directory
- centralize entry point to root index.js
- decouple tests from source files into a dedicated test/ directory
- extract examples to examples directory
- standardize TypeScript definitions in types directory
