+++ examples/example.js
@@ -1,5 +1,5 @@
 'use strict'
-const S = require('./FluentJSONSchema')
+const S = require('../FluentJSONSchema')
 const Ajv = require('ajv')
 
 const ROLES = {
