+++ lib/ArraySchema.js
