+++ lib/BaseSchema.js
