+++ lib/BooleanSchema.js
