+++ lib/IntegerSchema.js
