+++ lib/StringSchema.js
