+++ lib/schemas/basic.json
